==> 0_master_script.R <==
# Master script

source("WIT1_cleanup.R")
source("WIT1_analysis.R")
source("WIT2_cleanup.R")
source("WIT2_analysis.R")
source("WIT3_cleanup.R")
source("WIT3_analysis.R")
source("WIT4_cleanup.R")
source("WIT4_analysis.R")
